+++ pkg/stack/stack.go
@@ -14,7 +14,7 @@ type Stack struct {
 	head *frame
 	tail *frame
 	size int
-	mux  sync.Mutex
+	mux  sync.RWMutex
 }
 
 // frame represents an element of the stack. It contains
@@ -170,11 +170,17 @@ func (s *Stack) SweepPush(element interface{}) (interface{}, bool) {
 
 // Size returns the number of elements that a stack contains.
 func (s *Stack) Size() int {
+	s.mux.RLock()
+	defer s.mux.RUnlock()
+
 	return s.size
 }
 
 // Peek returns the element on top of the stack.
 func (s *Stack) Peek() interface{} {
+	s.mux.RLock()
+	defer s.mux.RUnlock()
+
 	if s.head == nil {
 		return nil
 	}