+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
 - Add option to allow pushing on a Stack when this is full
+
+### Changed
+- pkg/stack: Use RWMutex for concurrent reads
+
+### Fixed
+- pkg/stack: Fix data race conditions on Size and Peek
 
 ## [0.1.2] - 2017-03-05
 
@@ -17,7 +24,7 @@ All notable changes to this project will be documented in this file.
 - dev: Add pila.sh utilities to Docker image
 
 ### Changed
-- pilad: Show links of interest in `/` endpoint.
+- pilad: Show links of interest in `/` endpoint
 
 ### Fixed
 - Fix decoding bug when pushing a malformed payload